Pull out fmt_install.h and add fmt_install() to add custom specifiers
